Adds emoji restriction feature for non-admin users
