+++ apps/meteor/app/lib/server/methods/updateMessage.ts
@@ -9,6 +9,7 @@ import { canSendMessageAsync } from '../../../authorization/server/functions/can
 import { hasPermissionAsync } from '../../../authorization/server/functions/hasPermission';
 import { applyAirGappedRestrictionsValidation } from '../../../license/server/airGappedRestrictionsWrapper';
 import { settings } from '../../../settings/server';
+import { assertNoRestrictedEmojis } from '../lib/validateRestrictedEmojis';
 import { updateMessage } from '../functions/updateMessage';
 
 const allowedEditedFields = ['tshow', 'alias', 'attachments', 'avatar', 'emoji', 'msg', 'customFields', 'content', 'e2eMentions'];
@@ -86,27 +87,7 @@ export async function executeUpdateMessage(
 	}
 	await canSendMessageAsync(message.rid, { uid: user._id, username: user.username ?? undefined, ...user });
 
-	const restrictedEmojisString = settings.get<string>('Emoji_Restricted_For_Users');
-	if (restrictedEmojisString && message.msg) {
-		const hasManageEmojiPermission = await hasPermissionAsync(uid, 'manage-emoji');
-		if (!hasManageEmojiPermission) {
-			const restrictedEmojis = restrictedEmojisString.split(',').map((emoji) => emoji.trim()).filter(Boolean);
-			if (restrictedEmojis.length) {
-				const emojiRegex = /:([a-zA-Z0-9_+-]+):/g;
-				const matches = message.msg.matchAll(emojiRegex);
-				
-				for (const match of matches) {
-					const emojiName = match[1];
-					if (restrictedEmojis.includes(emojiName)) {
-						throw new Meteor.Error('error-emoji-restricted', `Emoji :${emojiName}: is restricted`, {
-							method: 'updateMessage',
-							emoji: emojiName,
-						});
-					}
-				}
-			}
-		}
-	}
+	await assertNoRestrictedEmojis(uid, message.msg, 'updateMessage');
 
 	message.u = originalMessage.u;
